update fix to parse cmd
