+++ cmd/bdjuno/main.go
@@ -3,16 +3,16 @@ package main
 import (
 	"github.com/cosmos/cosmos-sdk/simapp"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	parsecmd "github.com/forbole/bdjuno/v2/cmd/parse"
 	"github.com/forbole/juno/v3/cmd"
 	initcmd "github.com/forbole/juno/v3/cmd/init"
 	parsetypes "github.com/forbole/juno/v3/cmd/parse/types"
-	parsecmd "github.com/forbole/juno/v3/cmd/parse"
 
 	"github.com/forbole/juno/v3/modules/messages"
 
 	actionscmd "github.com/forbole/bdjuno/v2/cmd/actions"
-	startcmd "github.com/forbole/juno/v3/cmd/start"
 	migratecmd "github.com/forbole/juno/v3/cmd/migrate"
+	startcmd "github.com/forbole/juno/v3/cmd/start"
 
 	"github.com/forbole/bdjuno/v2/types/config"
