+++ cmd/parse/auth/cmd.go
@@ -1,12 +1,12 @@
 package auth
 
 import (
-	"github.com/forbole/juno/v3/cmd/parse"
+	parsecmdtypes "github.com/forbole/juno/v3/cmd/parse/types"
 	"github.com/spf13/cobra"
 )
 
 // NewAuthCmd returns the Cobra command that allows to fix all the things related to the x/auth module
-func NewAuthCmd(parseCfg *parse.Config) *cobra.Command {
+func NewAuthCmd(parseCfg *parsecmdtypes.Config) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "auth",
 		Short: "Fix things related to the x/auth module",
