+++ cmd/parse/staking/cmd.go
@@ -1,12 +1,12 @@
 package staking
 
 import (
-	"github.com/forbole/juno/v3/cmd/parse"
+	parsecmdtypes "github.com/forbole/juno/v3/cmd/parse/types"
 	"github.com/spf13/cobra"
 )
 
 // NewStakingCmd returns the Cobra command that allows to fix all the things related to the x/staking module
-func NewStakingCmd(parseConfig *parse.Config) *cobra.Command {
+func NewStakingCmd(parseConfig *parsecmdtypes.Config) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "staking",
 		Short: "Fix things related to the x/staking module",
